Migrate field.
