+++ crates/macros/src/common/mod.rs
@@ -1,5 +1,8 @@
 mod container;
 mod field;
+mod field_value;
 mod variant;
 
+pub use field::*;
+pub use field_value::*;
 pub use variant::*;
