+++ crates/macros/src/config/mod.rs
@@ -2,14 +2,13 @@
 pub mod config;
 pub mod config_type;
 pub mod setting;
-pub mod setting_type;
+pub mod setting_value;
 pub mod variant;
 
-use crate::common_schema::SerdeArgs;
+use crate::common::{Field, Variant};
+use crate::common_schema::ContainerSerdeArgs;
 use crate::config::config::{Config, ConfigArgs};
 use crate::config::config_type::ConfigType;
-use crate::config::setting::Setting;
-use crate::config::variant::Variant;
 use crate::utils::extract_common_attrs;
 use darling::FromDeriveInput;
 use proc_macro::TokenStream;
@@ -20,12 +19,12 @@ use syn::{parse_macro_input, Data, DeriveInput, Fields};
 pub fn macro_impl(item: TokenStream) -> TokenStream {
     let input: DeriveInput = parse_macro_input!(item);
     let args = ConfigArgs::from_derive_input(&input).expect("Failed to parse arguments.");
-    let serde_args = SerdeArgs::from_derive_input(&input).unwrap_or_default();
+    let serde_args = ContainerSerdeArgs::from_derive_input(&input).unwrap_or_default();
 
     let config_type = match &input.data {
         Data::Struct(data) => match &data.fields {
             Fields::Named(fields) => ConfigType::NamedStruct {
-                settings: fields.named.iter().map(Setting::from).collect::<Vec<_>>(),
+                settings: fields.named.iter().map(Field::from).collect::<Vec<_>>(),
                 fields,
             },
             Fields::Unnamed(_) => {
