+++ crates/macros/src/config_enum/variant.rs
@@ -1,4 +1,4 @@
-use crate::common::VariantSerdeArgs;
+use crate::common_schema::FieldSerdeArgs;
 use crate::utils::{extract_comment, extract_common_attrs, format_case, has_attr};
 use darling::FromAttributes;
 use proc_macro2::{Ident, TokenStream};
@@ -15,7 +15,7 @@ pub struct VariantArgs {
 
 pub struct Variant<'l> {
     pub args: VariantArgs,
-    pub serde_args: VariantSerdeArgs,
+    pub serde_args: FieldSerdeArgs,
     pub attrs: Vec<&'l Attribute>,
     pub name: &'l Ident,
     pub value: String,
@@ -24,7 +24,7 @@ pub struct Variant<'l> {
 impl<'l> Variant<'l> {
     pub fn from<'n>(variant: &'n NativeVariant, format: &str) -> Variant<'n> {
         let args = VariantArgs::from_attributes(&variant.attrs).unwrap_or_default();
-        let serde_args = VariantSerdeArgs::from_attributes(&variant.attrs).unwrap_or_default();
+        let serde_args = FieldSerdeArgs::from_attributes(&variant.attrs).unwrap_or_default();
 
         if args.fallback {
             if let Fields::Unnamed(fields) = &variant.fields {
